+++ src/functions/func_sequences.cpp
@@ -223,7 +223,7 @@ PlanIter_t fn_doc::codegen(
 {
   return new FnDocIterator ( sctx, loc, argv);
 }
-bool fn_doc::propagatesInputToOutput ( uint32_t ) const {
+bool fn_doc::propagatesInputToOutput ( ulong ) const {
   return false;
 }
 PlanIter_t fn_doc_available::codegen(
